+++ src/main/java/io/github/mendjoy/gymJourneyAPI/repository/WorkoutRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkoutRepository extends JpaRepository<Workout, Integer> {
+public interface WorkoutRepository extends JpaRepository<Workout, Long> {
 
     Page<Workout> findAllByUser(User user, Pageable pageable);
